joker/ai_core/pyproject.toml
[project]
name = "ai-core"
version = "0.1.0"
description = ""
authors = [
    {name = "xuews",email = "lenglengzailuguan@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"  # 修改这里，明确上限
dependencies = [
    "langchain[openai] (>=0.3.18,<0.4.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
